Move DW ogl endpoint to DW app
