+++ api/data_workspace/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class DataWorkspaceConfig(AppConfig):
-    name = 'data_workspace'
+    name = "data_workspace"
