+++ api/data_workspace/urls.py
@@ -7,6 +7,7 @@ app_name = "data_workspace"
 
 router = DefaultRouter()
 router.register("v1/licences", views.LicencesListDW, basename="dw-licences")
+router.register("v1/ogl", views.OpenGeneralLicenceListDW, basename="dw-ogl")
 
 urlpatterns = [
     path("", include(router.urls)),
