+++ api/data_workspace/views.py
@@ -1,8 +1,11 @@
 from django.db.models import Q
-from rest_framework import viewsets, pagination
+from rest_framework import viewsets
+from rest_framework.pagination import LimitOffsetPagination
 
 from api.cases.enums import CaseTypeEnum
 from api.core.authentication import DataWorkspaceOnlyAuthentication
+from api.open_general_licences.models import OpenGeneralLicence
+from api.open_general_licences.serializers import OpenGeneralLicenceSerializer
 from api.licences.models import Licence
 from api.licences.enums import LicenceStatus
 from api.licences.serializers.view_licence import LicenceListSerializer
@@ -13,7 +16,7 @@ from api.staticdata.statuses.models import CaseStatus
 class LicencesListDW(viewsets.ReadOnlyModelViewSet):
     authentication_classes = (DataWorkspaceOnlyAuthentication,)
     serializer_class = LicenceListSerializer
-    pagination_class = pagination.LimitOffsetPagination
+    pagination_class = LimitOffsetPagination
     non_active_states = CaseStatus.objects.filter(status=CaseStatusEnum.SURRENDERED)
 
     queryset = (
@@ -23,5 +26,18 @@ class LicencesListDW(viewsets.ReadOnlyModelViewSet):
             Q(case__status__in=non_active_states)
             | Q(case__case_type__id__in=CaseTypeEnum.OPEN_GENERAL_LICENCE_IDS)
             | Q(status=LicenceStatus.DRAFT)
-        ).order_by("created_at").reverse()
+        )
+        .order_by("created_at")
+        .reverse()
+    )
+
+
+class OpenGeneralLicenceListDW(viewsets.ReadOnlyModelViewSet):
+    authentication_classes = (DataWorkspaceOnlyAuthentication,)
+    serializer_class = OpenGeneralLicenceSerializer
+    pagination_class = LimitOffsetPagination
+    queryset = (
+        OpenGeneralLicence.objects.all()
+        .select_related("case_type")
+        .prefetch_related("countries", "control_list_entries")
     )
