+++ api/open_general_licences/tests/test_gets.py
@@ -129,24 +129,3 @@ class ExporterListTests(DataTestClient):
         self.assertEquals(response.status_code, status.HTTP_200_OK)
         self.assertEquals(len(response.json()["results"]), 1)
         self.assertEquals(len(response.json()["results"][0]["registrations"]), 2)
-
-
-class DataWorkspaceTests(DataTestClient):
-    def setUp(self):
-        super().setUp()
-        test_host = "http://testserver"
-        self.url = parse.urljoin(test_host, reverse("open_general_licences:ogl-dw-list"))
-
-    @override_settings(HAWK_AUTHENTICATION_ENABLED=True)
-    def test_dw_view_ogl_types(self):
-        sender = get_hawk_sender("GET", self.url, None, settings.HAWK_LITE_DATA_WORKSPACE_CREDENTIALS)
-        self.client.credentials(HTTP_HAWK_AUTHENTICATION=sender.request_header, CONTENT_TYPE="application/json")
-        response = self.client.get(self.url)
-        self.assertEquals(response.status_code, status.HTTP_200_OK)
-
-    @override_settings(HAWK_AUTHENTICATION_ENABLED=True)
-    def test_dw_view_ogl_fail_incorrect_hawk_key(self):
-        sender = get_hawk_sender("GET", self.url, None, "internal-frontend")
-        self.client.credentials(HTTP_HAWK_AUTHENTICATION=sender.request_header, CONTENT_TYPE="application/json")
-        with self.assertRaises(mohawk.exc.HawkFail):
-            self.client.get(self.url)
